requirements.txt
anyio==4.4.0
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
filelock==3.15.4
Flask==2.0.1
fsspec==2024.6.1
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-texttospeech==2.8.0
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.1
pillow==10.4.0
pip==24.0
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rsa==4.9
sacremoses==0.1.1
sentencepiece==0.1.96
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tbb==2021.13.0
tokenizers==0.13.3
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.27.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==2.0.1
wheel==0.43.0
python-dotenv
supervisor
pyTelegramBotAPI
